+++ crates/diesel_models/src/payment_intent.rs
@@ -72,6 +72,7 @@ pub struct PaymentIntent {
     pub tax_on_surcharge: Option<MinorUnit>,
     // TODO: change this to global id
     pub id: common_utils::id_type::PaymentId,
+    pub skip_external_tax_calculation: Option<bool>,
 }
 
 #[cfg(all(any(feature = "v1", feature = "v2"), not(feature = "payment_v2")))]
@@ -135,6 +136,7 @@ pub struct PaymentIntent {
     pub shipping_cost: Option<MinorUnit>,
     pub organization_id: common_utils::id_type::OrganizationId,
     pub tax_details: Option<TaxDetails>,
+    pub skip_external_tax_calculation: Option<bool>,
 }
 
 #[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize, diesel::AsExpression)]
@@ -219,6 +221,7 @@ pub struct PaymentIntentNew {
     pub tax_on_surcharge: Option<MinorUnit>,
     // TODO: change this to global id
     pub id: common_utils::id_type::PaymentId,
+    pub skip_external_tax_calculation: Option<bool>,
 }
 
 #[cfg(all(any(feature = "v1", feature = "v2"), not(feature = "payment_v2")))]
@@ -283,6 +286,7 @@ pub struct PaymentIntentNew {
     pub shipping_cost: Option<MinorUnit>,
     pub organization_id: common_utils::id_type::OrganizationId,
     pub tax_details: Option<TaxDetails>,
+    pub skip_external_tax_calculation: Option<bool>,
 }
 
 #[cfg(all(feature = "v2", feature = "payment_v2"))]