+++ crates/hyperswitch_domain_models/src/payments/payment_attempt.rs
@@ -112,6 +112,7 @@ pub trait PaymentAttemptInterface {
         payment_method_type: Option<Vec<storage_enums::PaymentMethodType>>,
         authentication_type: Option<Vec<storage_enums::AuthenticationType>>,
         merchant_connector_id: Option<Vec<id_type::MerchantConnectorAccountId>>,
+        profile_id_list: Option<Vec<id_type::ProfileId>>,
         storage_scheme: storage_enums::MerchantStorageScheme,
     ) -> error_stack::Result<i64, errors::StorageError>;
 }
@@ -549,6 +550,7 @@ impl behaviour::Conversion for PaymentIntent {
             organization_id: self.organization_id,
             shipping_cost: self.shipping_cost,
             tax_details: self.tax_details,
+            skip_external_tax_calculation: self.skip_external_tax_calculation,
         })
     }
     async fn convert_back(
@@ -633,6 +635,7 @@ impl behaviour::Conversion for PaymentIntent {
                 tax_on_surcharge: storage_model.tax_on_surcharge,
                 shipping_cost: storage_model.shipping_cost,
                 tax_details: storage_model.tax_details,
+                skip_external_tax_calculation: storage_model.skip_external_tax_calculation,
             })
         }
         .await
@@ -693,6 +696,7 @@ impl behaviour::Conversion for PaymentIntent {
             organization_id: self.organization_id,
             shipping_cost: self.shipping_cost,
             tax_details: self.tax_details,
+            skip_external_tax_calculation: self.skip_external_tax_calculation,
         })
     }
 }
@@ -756,6 +760,7 @@ impl behaviour::Conversion for PaymentIntent {
             organization_id: self.organization_id,
             shipping_cost: self.shipping_cost,
             tax_details: self.tax_details,
+            skip_external_tax_calculation: self.skip_external_tax_calculation,
         })
     }
 
@@ -842,6 +847,7 @@ impl behaviour::Conversion for PaymentIntent {
                     .await?,
                 is_payment_processor_token_flow: storage_model.is_payment_processor_token_flow,
                 organization_id: storage_model.organization_id,
+                skip_external_tax_calculation: storage_model.skip_external_tax_calculation,
             })
         }
         .await
@@ -903,6 +909,7 @@ impl behaviour::Conversion for PaymentIntent {
             organization_id: self.organization_id,
             shipping_cost: self.shipping_cost,
             tax_details: self.tax_details,
+            skip_external_tax_calculation: self.skip_external_tax_calculation,
         })
     }
 }